Revert "mm-video: vdec: Support adaptive playback mode"
This reverts commit d3b9062f2fb66a14338db52f4c117b4231725741.

This is the first of 3 reverts to disable adaptive playback and use the
legacy smooth streaming codepath instead.

Adaptive playback to a TextureView requires the OpenGL stack to handle the case
where the buffer's width and height may change during playback. The Qualcomm
Userspace OpenGL blobs we are using cannot correctly handle this.

Disabling adaptive playback solves the video corruption problem mentioned here:
https://github.com/google/ExoPlayer/issues/58
The fix to gralloc mentioned in that thread cannot be used because there
also has to be a corresponding change to the Qualcomm Userspace OpenGL
blobs, but AFAIK no such blobs exist that also support the old Adreno
220 GPU.